+++ source/infrastructure/J3ByteSource.m
@@ -8,11 +8,11 @@
 
 @implementation J3ByteSource
 
-+ (void) ensureBytesReadFromSource: (id <J3ByteSource>) byteSource intoBuffer: (uint8_t *) buffer ofLength: (unsigned) length;
++ (void) ensureBytesReadFromSource: (id <J3ByteSource>) byteSource intoBuffer: (uint8_t *) buffer ofLength: (size_t) length;
 {
   for (unsigned i = 0; i < length; i++)
   {
-    NSData *bytes = [byteSource readUpToLength: 1];
+    NSData *bytes = [byteSource readUpToLength: (size_t) 1];
     if ([bytes length] != 0) 
       buffer[i] = ((uint8_t *) [bytes bytes])[0];
     else
