+++ source/infrastructure/telnet/J3TelnetConnection.m
@@ -8,7 +8,7 @@
 #import "J3Socket.h"
 #import "J3TelnetConnection.h"
 
-#define TELNET_READ_BUFFER_SIZE 512
+#define TELNET_READ_BUFFER_SIZE (size_t) 512
 
 #pragma mark -
